Don't stop ESC propagation unless dropdown is open
In reference to issue #3096
https://github.com/angular-ui/bootstrap/issues/3096
If the escape key should have other functionality in the context of a form (for example, if datepicker is on a modal) it makes sense to prevent this from happening if escape is used to close the dropdown.  If the dropdown is closed, however, the event should be allowed to propagate.